+++ src/blockly/components/ActionContainer.js
@@ -4,11 +4,12 @@
  *
  * @flow
  */
-import { colors, blockNames } from 'blockly/constants';
+import * as colors from 'constants/colors';
+import { ACTION_CONTAINER } from 'constants/actions';
 
 const mainInput = 'program';
 
-Blockly.Blocks[blockNames.ACTION_CONTAINER] = {
+Blockly.Blocks[ACTION_CONTAINER] = {
   init: function init() {
     this.appendDummyInput()
         .appendField(Blockly.Msg.ACTION_CONTAINER_TITLE);
@@ -21,5 +22,5 @@ Blockly.Blocks[blockNames.ACTION_CONTAINER] = {
   },
 };
 
-Blockly.JavaScript[blockNames.ACTION_CONTAINER] =
+Blockly.JavaScript[ACTION_CONTAINER] =
   (block: Object) => (Blockly.JavaScript.statementToCode(block, mainInput));